Added a minor knob

+++ src/learning_engine_featurewise.cc
@@ -40,6 +40,7 @@ namespace knob
 	extern bool 			le_featurewise_enable_score_plot;
 	extern vector<int32_t> le_featurewise_plot_actions;
 	extern std::string 	le_featurewise_plot_file_name;
+	extern bool 			le_featurewise_remove_plot_script;
 }
 
 void LearningEngineFeaturewise::init_knobs()
@@ -473,6 +474,9 @@ void LearningEngineFeaturewise::plot_scores()
 	std::string cmd = "gnuplot " + std::string(script_file);
 	system(cmd.c_str());
 
-	std::string cmd2 = "rm " + std::string(script_file);
-	system(cmd2.c_str());
+	if(knob::le_featurewise_remove_plot_script)
+	{
+		std::string cmd2 = "rm " + std::string(script_file);
+		system(cmd2.c_str());
+	}
 }
